Criação do DTO Main para puxar dados da resposta da API

+++ src/main/java/model/dto/Weather/Main.java
@@ -1,13 +1,13 @@
 package model.dto.Weather;
 
 public class Main {
-    private int temp,feels_like,temp_min,temp_max;
+    private double temp,feels_like,temp_min,temp_max;
 
-    public int getTemp() {return temp;}
+    public double getTemp() {return temp;}
 
-    public int getFeels_like() {return feels_like;}
+    public double getFeels_like() {return feels_like;}
 
-    public int getTemp_min() {return temp_min;}
+    public double getTemp_min() {return temp_min;}
 
-    public int getTemp_max() {return temp_max;}
+    public double getTemp_max() {return temp_max;}
 }
